refactor: 순수 모델 삭제 - JPA 엔티티로 통합
