--- src/main/kotlin/com/sendy/domain/account/TransactionHistoryRepository.kt
@@ -1,5 +1,0 @@
-package com.sendy.domain.account
-
-interface TransactionHistoryRepository {
-    fun save(domain: TransactionHistory)
-}
