--- src/main/kotlin/com/sendy/infrastructure/persistence/account/AccountJpaRepository.kt
@@ -1,7 +1,0 @@
-package com.sendy.infrastructure.persistence.account
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface AccountJpaRepository : JpaRepository<AccountEntity, Long> {
-    fun findByUserId(userId: Long): List<AccountEntity>
-}
