+++ src/main/kotlin/com/sendy/infrastructure/persistence/account/AccountRepository.kt
@@ -1,0 +1,9 @@
+package com.sendy.infrastructure.persistence.account
+
+import com.sendy.domain.account.AccountEntity
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface AccountRepository : JpaRepository<AccountEntity, Long> {
+    fun findByUserId(userId: Long): List<AccountEntity>
+    fun findByUserIdAndAccountNumber(userId: Long, accountNumber: String): AccountEntity?
+}
