--- src/main/kotlin/com/sendy/infrastructure/persistence/account/TransactionHistoryJpaRepository.kt
@@ -1,9 +1,0 @@
-package com.sendy.infrastructure.persistence.account
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface TransactionHistoryJpaRepository :
-    JpaRepository<
-        TransactionHistoryJpaEntity,
-        Long,
-    >
